requirements.txt
paho-mqtt
gpxpy
numpy
